fix: created fixture-unit to separate logic between unit and solo
