+++ packages/network/tests/signer/signers/vechain-private-key-signer/vechain-private-key-signer.unit.test.ts
@@ -22,7 +22,7 @@ import {
     eip712TestCases,
     populateCallTestCases,
     populateCallTestCasesAccount
-} from './fixture';
+} from './fixture-unit';
 
 /**
  * VeChain base signer tests
